requirements.txt
promptflow-tools
promptflow>=1.8.0
instructor==1.6.3
pydantic==2.9.2
tenacity==9.0.0
google.generativeai==0.8.3
typer
pre-commit
langfuse==2.53.1
uvicorn
streamlit
gradio>=5.0.2
anthropic
